move backend ssa pass registration into backend modules
