+++ src/ninetoothed/backends/tvm.py
@@ -7,6 +7,8 @@ not through kernel-specialized lowerer branches.
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING, Any, Mapping
+
 from ninetoothed.backends.base import (
     Backend,
     BackendArtifact,
@@ -16,6 +18,9 @@ from ninetoothed.backends.base import (
 )
 from ninetoothed.backends.ssa_unified import lower_unified_ssa_artifact
 from ninetoothed.ir import KernelIR
+
+if TYPE_CHECKING:
+    from ninetoothed.ssa_passes import SSAPassRegistry
 
 
 class TvmBackend(Backend):
@@ -35,3 +40,46 @@ class TvmBackend(Backend):
         self, kernel: KernelIR, options: BackendOptions | None = None
     ) -> BackendArtifact:
         return lower_unified_ssa_artifact(kernel, self.name)
+
+
+def _scheduled_tir_loop_policy(schedule: Mapping[str, Any]) -> Mapping[str, Any]:
+    return {
+        "passes": ("block-tiling", "loop-reorder"),
+        "lowering": "scheduled-tir-loops",
+        "tile": dict(schedule.get("tile", {})),
+    }
+
+
+def _generic_linear_or_reduction_policy(
+    schedule: Mapping[str, Any],
+) -> Mapping[str, Any]:
+    if schedule.get("granularity") == "parallel-reduction":
+        return {
+            "passes": ("tree-reduction",),
+            "lowering": "ssa-reduction-scf-loop",
+        }
+    return {
+        "passes": ("coalesced-linear-indexing",),
+        "lowering": "ssa-operation-linear-emission",
+    }
+
+
+def register_ssa_passes(registry: "SSAPassRegistry") -> None:
+    from ninetoothed.ssa_passes import OptimizeSchedulePass
+
+    class TvmOptimizeSchedulePass(OptimizeSchedulePass):
+        name = "ssa.tvm.optimize_schedule"
+        supported_backends = (BackendName.TVM,)
+
+        def optimization_policy(
+            self,
+            backend: BackendName,
+            analysis: Mapping[str, Any],
+            schedule: Mapping[str, Any],
+        ) -> Mapping[str, Any]:
+            del backend, analysis
+            if schedule.get("granularity") == "blocked-linalg":
+                return _scheduled_tir_loop_policy(schedule)
+            return _generic_linear_or_reduction_policy(schedule)
+
+    registry.register(TvmOptimizeSchedulePass, tags=("optimization", "tvm"))
